telegram-bot-project/requirements.txt
python-telegram-bot==20.0
requests==2.26.0
logging==0.5.1.2